+++ app/src/main/java/org/simple/clinic/main/TheActivityEffect.kt
@@ -1,7 +1,5 @@
 package org.simple.clinic.main
 
-import java.time.Instant
-
 sealed class TheActivityEffect
 
 object LoadAppLockInfo : TheActivityEffect()
@@ -9,8 +7,6 @@ object LoadAppLockInfo : TheActivityEffect()
 object ClearLockAfterTimestamp : TheActivityEffect()
 
 object ShowAppLockScreen : TheActivityEffect()
-
-data class UpdateLockTimestamp(val lockAt: Instant) : TheActivityEffect()
 
 object ListenForUserVerifications : TheActivityEffect()