+++ app/src/test/java/org/simple/clinic/login/applock/AppLockScreenLogicTest.kt
@@ -1,17 +1,18 @@
 package org.simple.clinic.login.applock
 
-import com.f2prateek.rx.preferences2.Preference
+import com.google.common.truth.Truth.assertThat
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
-import dagger.Lazy
 import io.reactivex.rxkotlin.ofType
 import io.reactivex.subjects.PublishSubject
 import org.junit.After
 import org.junit.Rule
 import org.junit.Test
 import org.simple.clinic.TestData
+import org.simple.clinic.storage.MemoryValue
+import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.scheduler.TestSchedulersProvider
 import org.simple.clinic.widgets.UiEvent
@@ -26,7 +27,7 @@ class AppLockScreenLogicTest {
 
   private val ui = mock<AppLockScreenUi>()
   private val uiActions = mock<AppLockUiActions>()
-  private val lastUnlockTimestamp = mock<Preference<Instant>>()
+  private val lockAfterTimestampValue = MemoryValue(Optional.empty<Instant>())
 
   private val loggedInUser = TestData.loggedInUser(
       uuid = UUID.fromString("cdb08a78-7bae-44f4-9bb9-40257be58aa4"),
@@ -48,13 +49,15 @@ class AppLockScreenLogicTest {
 
   @Test
   fun `when PIN is authenticated, the last-unlock-timestamp should be updated and then the app should be unlocked`() {
+    // given
+    lockAfterTimestampValue.set(Optional.of(Instant.parse("2018-01-01T00:00:00Z")))
+
     // when
     setupController()
     uiEvents.onNext(AppLockPinAuthenticated)
 
     // then
-    verify(lastUnlockTimestamp).delete()
-
+    assertThat(lockAfterTimestampValue.hasValue).isFalse()
     verify(ui, times(2)).setUserFullName(loggedInUser.fullName)
     verify(ui).setFacilityName(facility.name)
     verify(uiActions).restorePreviousScreen()
@@ -104,10 +107,10 @@ class AppLockScreenLogicTest {
 
   private fun setupController() {
     val effectHandler = AppLockEffectHandler(
-        currentUser = Lazy { loggedInUser },
-        currentFacility = Lazy { facility },
-        lockAfterTimestamp = lastUnlockTimestamp,
+        currentUser = { loggedInUser },
+        currentFacility = { facility },
         schedulersProvider = TestSchedulersProvider.trampoline(),
+        lockAfterTimestampValue = lockAfterTimestampValue,
         uiActions = uiActions
     )